# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QA Training Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/progress:
    get:
      operationId: getUserProgress
      tags: [users]
      summary: Get user learning progress
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProgress"
...
  /exercises:
    get:
      operationId: getExercises
      tags: [exercises]
      summary: List exercises
      parameters:
        - name: lessonId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
  /exercises/{id}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get exercise details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseDetail"
  /submissions:
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit an exercise
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionRequest"
      responses:
        "201":
          description: Submission evaluated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResult"
  /submissions/user/{userId}:
    get:
      operationId: getUserSubmissions
      tags: [submissions]
      summary: Get submissions for a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
...
  /test-cases:
    get:
      operationId: getTestCases
      tags: [testCases]
      summary: List test cases
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: exerciseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of test cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestCase"
    post:
      operationId: createTestCase
      tags: [testCases]
      summary: Submit a test case
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestCaseRequest"
      responses:
        "201":
          description: Test case submitted and evaluated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCaseResult"
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get global leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /badges/user/{userId}:
    get:
      operationId: getUserBadges
      tags: [leaderboard]
      summary: Get badges for a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User badges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Badge"
  /certificates/user/{userId}:
    get:
      operationId: getUserCertificates
      tags: [certificates]
      summary: Get certificates for a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certificate"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [student, instructor, admin]
        learningPath:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        totalScore:
          type: integer
        level:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - totalScore
        - level
        - createdAt
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        learningPath:
          type: string
        avatarUrl:
          type: string
    UserProgress:
      type: object
      properties:
        userId:
          type: integer
        enrolledCourses:
          type: integer
        completedCourses:
          type: integer
        completedLessons:
          type: integer
        totalLessons:
          type: integer
        totalScore:
          type: integer
        submissionsCount:
          type: integer
        bugReportsCount:
          type: integer
        testCasesCount:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required:
        - userId
        - enrolledCourses
        - completedCourses
        - completedLessons
        - totalLessons
        - totalScore
        - submissionsCount
        - bugReportsCount
        - testCasesCount
        - recentActivity
    ActivityItem:
      type: object
      properties:
        type:
          type: string
        description:
          type: string
        score:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - type
        - description
        - createdAt
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
        lessonId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [write-test-case, report-bug, write-automation, test-api, create-test-plan, quiz]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        points:
          type: integer
        description:
          type: string
      required:
        - id
        - lessonId
        - title
        - type
        - difficulty
        - points
        - description
    ExerciseDetail:
      allOf:
        - $ref: "#/components/schemas/Exercise"
        - type: object
          properties:
            instructions:
              type: string
            template:
              type: string
              nullable: true
            hints:
              type: array
              items:
                type: string
            demoAppId:
              type: integer
              nullable: true
          required:
            - instructions
            - hints
    CreateSubmissionRequest:
      type: object
      properties:
        exerciseId:
          type: integer
        content:
          type: string
        language:
          type: string
          nullable: true
      required:
        - exerciseId
        - content
    Submission:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        exerciseId:
          type: integer
        exerciseTitle:
          type: string
        content:
          type: string
        score:
          type: integer
        maxScore:
          type: integer
        feedback:
          type: string
        suggestions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, evaluated, passed, failed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - exerciseId
        - exerciseTitle
        - content
        - score
        - maxScore
        - feedback
        - suggestions
        - status
        - createdAt
    SubmissionResult:
      allOf:
        - $ref: "#/components/schemas/Submission"
        - type: object
          properties:
            pointsEarned:
              type: integer
            badgeEarned:
              $ref: "#/components/schemas/Badge"
              nullable: true
          required:
            - pointsEarned
...
    CreateTestCaseRequest:
      type: object
      properties:
        exerciseId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        preconditions:
          type: string
        testSteps:
          type: string
        expectedResult:
          type: string
        testType:
          type: string
          enum: [functional, regression, smoke, sanity, integration, performance]
        priority:
          type: string
          enum: [critical, high, medium, low]
      required:
        - title
        - description
        - preconditions
        - testSteps
        - expectedResult
        - testType
        - priority
    TestCase:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        exerciseId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        preconditions:
          type: string
        testSteps:
          type: string
        expectedResult:
          type: string
        testType:
          type: string
          enum: [functional, regression, smoke, sanity, integration, performance]
        priority:
          type: string
          enum: [critical, high, medium, low]
        score:
          type: integer
          nullable: true
        feedback:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - title
        - description
        - preconditions
        - testSteps
        - expectedResult
        - testType
        - priority
        - createdAt
    TestCaseResult:
      type: object
      properties:
        testCase:
          $ref: "#/components/schemas/TestCase"
        score:
          type: integer
        feedback:
          type: string
        suggestions:
          type: array
          items:
            type: string
      required:
        - testCase
        - score
        - feedback
        - suggestions
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        totalScore:
          type: integer
        level:
          type: integer
        badgesCount:
          type: integer
        completedCourses:
          type: integer
      required:
        - rank
        - userId
        - name
        - totalScore
        - level
        - badgesCount
        - completedCourses
    Badge:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        earnedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - name
        - description
        - icon
        - color
    Certificate:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        courseTitle:
          type: string
        issuedAt:
          type: string
          format: date-time
        certificateUrl:
          type: string
      required:
        - id
        - userId
        - courseId
        - courseTitle
        - issuedAt
        - certificateUrl
...