# lib/api-zod/src/generated/types/testCaseResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QA Training Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TestCase } from "./testCase";

export interface TestCaseResult {
  testCase: TestCase;
  score: number;
  feedback: string;
  suggestions: string[];
}